separate Client class
