+++ lib/uni_sender/version.rb
@@ -1,3 +1,3 @@
 module UniSender
-  VERSION = "0.0.7"
+  VERSION = '0.0.7'.freeze
 end
